@@ -706,9 +706,9 @@
 		"await-notify": "^1.0.1",
 		"crc-32": "^1.2.2",
 		"fs-extra": "^11.2.0",
-		"npm-check-updates": "^17.1.9",
+		"npm-check-updates": "^17.1.10",
 		"original-fs": "^1.2.0",
-		"pinia": "^2.2.5",
+		"pinia": "^2.2.6",
 		"socket.io": "^4.8.1",
 		"uuid": "^11.0.2",
 		"vee-validate": "^4.14.6",
@@ -721,8 +721,7 @@
 		"@types/archiver": "^6.0.3",
 		"@types/bootstrap": "^5.2.10",
 		"@types/fs-extra": "^11.0.4",
-		"@types/jest": "^29.5.14",
-		"@types/node": "^22.8.4",
+		"@types/node": "^22.8.7",
 		"@types/power-assert": "^1.5.12",
 		"@types/psd": "^3.4.3",
 		"@types/vscode": "^1.95.0",
@@ -735,31 +734,29 @@
 		"esbuild": "^0.24.0",
 		"file-loader": "^6.2.0",
 		"image-size": "^1.1.1",
-		"jest": "^29.7.0",
 		"p-queue": "^8.0.1",
 		"rimraf": "^6.0.1",
 		"terser": "^5.36.0",
-		"ts-jest": "^29.2.5",
-		"ts-jest-resolver": "^2.0.1",
 		"ts-loader": "^9.5.1",
 		"tsx": "^4.19.2",
 		"typescript": "^5.6.3",
 		"vite": "^5.4.10",
+		"vitest": "^2.1.4",
 		"vscode-languageserver-types": "^3.17.5",
 		"vue-loader": "^17.4.2",
-		"webpack": "^5.95.0",
+		"webpack": "^5.96.1",
 		"webpack-cli": "^5.1.4"
 	},
 	"optionalDependencies": {
 		"bufferutil": "^4.0.8",
-		"utf-8-validate": "^6.0.4"
+		"utf-8-validate": "^6.0.5"
 	},
 	"scripts": {
 		"build": "tsx build.ts",
 		"watch": "tsx watch build.ts --watch",
 		"md2json": "tsx src/md2json.ts && npm run build:types",
-		"build:types": "tsc -P tsconfig.types.json",
-		"vue_setting_preview": "vite",
+		"build:types": "tsc -P ./src/tsconfig.types.json",
+		"vue_setting_preview": "vite --config ./views/vite.config.mts",
 		"npm_check_updates": "ncu -u --target minor && npm update && npm run build && npm run ncu_lsp && ncu -g",
 		"ncu_lsp": "cd server && ncu -u --target minor && npm update && npm outdated && cd ..",
 		"postinstall": "cd server && npm install && cd ..",
@@ -772,12 +769,7 @@
 		"wp:pro_es": "tsx build.ts --production",
 		"wp:pro": "webpack --config ./src/webpack.config.js --mode production",
 		"package": "webpack --config ./src/webpack.config.js --mode production --devtool hidden-source-map",
-		"test": "jest"
-	},
-	"jest": {
-		"transform": {
-			"^.+\\.tsx?$": "ts-jest"
-		}
+		"test": "vitest"
 	},
 	"repository": {
 		"type": "git",
